Add logic for checking a round's win state
playRound has a rather ugly looking conditional check for if the player
wins, loses, or draws. I don't think a switch case applies here.
Perhaps better text formatting will work, or give it its own function.
The purpose of playRound is to be used as many times as desired by
the game() function and keep score for each round played.