Create function for asking for human choice
This does not check if the prompt input is strictly rock, paper, or
scissors, so functionality depends entirely on user trust. This is of
course not a good idea, but we're keeping it simple since this is just
groundwork for making a proper UI later instead of playing on console.